+++ src/main/java/fr/inrae/act/bagap/chloe/window/counting/QuantitativeCounting.java
@@ -2,9 +2,9 @@ package fr.inrae.act.bagap.chloe.window.counting;
 
 public class QuantitativeCounting extends Counting implements QuantitativeCountingInterface {
 	
-	private double sum, squareSum, minimum, maximum;
+	private float sum, squareSum, minimum, maximum;
 	
-	public QuantitativeCounting(double theoriticalSize) {
+	public QuantitativeCounting(float theoriticalSize) {
 		super(theoriticalSize);
 	}
 	
@@ -20,7 +20,7 @@ public class QuantitativeCounting extends Counting implements QuantitativeCounti
 	 * 7 : maximum des valeurs
 	 */
 	@Override
-	public void doSetCounts(double[] counts){
+	public void doSetCounts(float[] counts){
 		sum = counts[4];
 		squareSum = counts[5];
 		minimum = counts[6];
@@ -28,35 +28,35 @@ public class QuantitativeCounting extends Counting implements QuantitativeCounti
 	}
 
 	@Override
-	public double average() {
+	public float average() {
 		return (float) (sum / validValues());
 	}
 
 	@Override
-	public double sum() {
+	public float sum() {
 		return sum;
 	}
 	
 	@Override
-	public double squareSum() {
+	public float squareSum() {
 		return squareSum;
 	}
 	
 	@Override
-	public double standardDeviation() {
+	public float standardDeviation() {
 		if(((squareSum / validValues()) - Math.pow(sum / validValues(), 2)) < 0){
 			return 0;
 		}
-		return Math.sqrt((squareSum / validValues()) - Math.pow(sum / validValues(), 2));
+		return (float) Math.sqrt((squareSum / validValues()) - Math.pow(sum / validValues(), 2));
 	}
 	
 	@Override
-	public double minimum(){
+	public float minimum(){
 		return minimum;
 	}
 	
 	@Override
-	public double maximum(){
+	public float maximum(){
 		return maximum;
 	}
